Remove AndroidX Startup and Screenshot testing dependencies
This commit removes the AndroidX Startup library as a dependency from the `app` and `core/database` modules.
Additionally, it temporarily comments out the screenshot testing setup, including the plugin application in `ComposeConventionPlugin.kt`, the `android.experimental.enableScreenshotTest` property, and related dependencies. This is likely a temporary measure pending an update or resolution of an issue with the screenshot testing functionality.